refactor(auth): isolate browser sign-in persistence
