--- packages/control-plane/src/auth/immutable-provider-identity.ts
@@ -1,379 +1,0 @@
-import type { ProviderCredentialInput } from "./provider-credential";
-import type { ProviderCodeExchangeResult, VerifiedProviderIdentity } from "./providers/types";
-import type { SignInProvider } from "./sign-in-provider";
-import { isUniqueConstraintError } from "../db/errors";
-import type { SqlDatabase, SqlStatement } from "../db/sql-database";
-
-const CANONICAL_ISSUERS: Readonly<Record<SignInProvider, string>> = {
-  github: "https://github.com",
-  google: "https://accounts.google.com",
-};
-const MAX_RESOLUTION_ATTEMPTS = 3;
-const MAX_VERIFIED_EMAIL_CLAIMS = 1_000;
-
-export interface ResolvedImmutableProviderIdentity {
-  readonly userId: string;
-  readonly providerIdentityId: string;
-  readonly isNewUser: boolean;
-  readonly collisionCount: number;
-}
-
-export interface ImmutableProviderIdentityDependencies {
-  readonly clock: { now(): number };
-  readonly idGenerator: { generate(): string };
-  readonly providerCredentialStore: ProviderCredentialStorePort;
-}
-
-export interface ProviderCredentialStorePort {
-  prepareInitialInsert(
-    providerIdentityId: string,
-    credential: ProviderCredentialInput,
-    updatedAt: number
-  ): Promise<SqlStatement>;
-  prepareSignInUpsert(
-    providerIdentityId: string,
-    credential: ProviderCredentialInput,
-    updatedAt: number
-  ): Promise<SqlStatement>;
-  isSignInVersionConflict(error: unknown): boolean;
-}
-
-export class InvalidProviderIdentityEvidenceError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "InvalidProviderIdentityEvidenceError";
-  }
-}
-
-export class AccountLinkRequiredError extends Error {
-  constructor(readonly collisionCount: number) {
-    super("This verified identity requires explicit account linking");
-    this.name = "AccountLinkRequiredError";
-  }
-}
-
-export class ProviderIdentityAdapterMismatchError extends Error {
-  constructor() {
-    super("Stored provider identity does not match the authenticating adapter");
-    this.name = "ProviderIdentityAdapterMismatchError";
-  }
-}
-
-interface EmailClaimRow {
-  email: string;
-  user_id: string;
-  source_kind: "legacy_canonical" | "provider_verified" | "trusted_bot_attribution";
-}
-
-interface IdentityRow {
-  id: string;
-  user_id: string;
-  provider: string;
-}
-
-interface NormalizedIdentityEvidence {
-  readonly provider: SignInProvider;
-  readonly issuer: string;
-  readonly subject: string;
-  readonly login: string | null;
-  readonly displayName: string | null;
-  readonly avatarUrl: string | null;
-  readonly verifiedEmails: readonly string[];
-  readonly primaryEmail: string | null;
-}
-
-function normalizeOptional(value: string | undefined): string | null {
-  const normalized = value?.trim();
-  return normalized ? normalized : null;
-}
-
-function normalizeIdentityEvidence(identity: VerifiedProviderIdentity): NormalizedIdentityEvidence {
-  if (identity.issuer !== CANONICAL_ISSUERS[identity.provider]) {
-    throw new InvalidProviderIdentityEvidenceError(
-      "Provider identity issuer is not the configured canonical issuer"
-    );
-  }
-  if (identity.subject.length === 0) {
-    throw new InvalidProviderIdentityEvidenceError("Provider identity subject is empty");
-  }
-
-  const verifiedEmails = [
-    ...new Set(identity.verifiedEmails.map((email) => email.trim().toLowerCase()).filter(Boolean)),
-  ];
-  if (verifiedEmails.length > MAX_VERIFIED_EMAIL_CLAIMS) {
-    throw new InvalidProviderIdentityEvidenceError(
-      "Provider identity has too many verified email claims"
-    );
-  }
-  const primaryEmail = identity.primaryEmail?.trim().toLowerCase() || null;
-  if (primaryEmail !== null && !verifiedEmails.includes(primaryEmail)) {
-    throw new InvalidProviderIdentityEvidenceError(
-      "Primary display email is not provider-verified"
-    );
-  }
-
-  return {
-    provider: identity.provider,
-    issuer: identity.issuer,
-    subject: identity.subject,
-    login: normalizeOptional(identity.login),
-    displayName: normalizeOptional(identity.displayName),
-    avatarUrl: normalizeOptional(identity.avatarUrl),
-    verifiedEmails,
-    primaryEmail,
-  };
-}
-
-function requireGeneratedId(value: string, kind: string): string {
-  if (value.length === 0) {
-    throw new Error(`Provider identity ${kind} generator returned an invalid id`);
-  }
-  return value;
-}
-
-export class ImmutableProviderIdentityService {
-  constructor(
-    private readonly db: SqlDatabase,
-    private readonly dependencies: ImmutableProviderIdentityDependencies
-  ) {}
-
-  async resolve(
-    signIn: ProviderCodeExchangeResult<SignInProvider>
-  ): Promise<ResolvedImmutableProviderIdentity> {
-    const identity = normalizeIdentityEvidence(signIn.identity);
-    const credential = signIn.credential;
-
-    for (let attempt = 1; attempt <= MAX_RESOLUTION_ATTEMPTS; attempt += 1) {
-      const existing = await this.findIdentity(identity.issuer, identity.subject);
-      if (existing) {
-        try {
-          return await this.refreshExisting(existing, identity, credential);
-        } catch (error) {
-          if (
-            attempt === MAX_RESOLUTION_ATTEMPTS ||
-            (!isUniqueConstraintError(error) &&
-              !this.dependencies.providerCredentialStore.isSignInVersionConflict(error))
-          ) {
-            throw error;
-          }
-          continue;
-        }
-      }
-
-      const conflictingEmails = await this.findConflictingEmails(identity.verifiedEmails, null);
-      if (conflictingEmails.length > 0) {
-        throw new AccountLinkRequiredError(conflictingEmails.length);
-      }
-
-      try {
-        return await this.createIdentity(identity, credential);
-      } catch (error) {
-        if (!isUniqueConstraintError(error) || attempt === MAX_RESOLUTION_ATTEMPTS) {
-          throw error;
-        }
-      }
-    }
-
-    throw new Error("Provider identity resolution exhausted its retry budget");
-  }
-
-  private async createIdentity(
-    identity: NormalizedIdentityEvidence,
-    credential: ProviderCredentialInput | null
-  ): Promise<ResolvedImmutableProviderIdentity> {
-    const now = this.dependencies.clock.now();
-    const userId = requireGeneratedId(this.dependencies.idGenerator.generate(), "user id");
-    const providerIdentityId = requireGeneratedId(
-      this.dependencies.idGenerator.generate(),
-      "identity id"
-    );
-
-    const statements: SqlStatement[] = [
-      this.db
-        .prepare(
-          `INSERT INTO users (
-             id, display_name, email, avatar_url, created_at, updated_at
-           ) VALUES (?, ?, ?, ?, ?, ?)`
-        )
-        .bind(userId, identity.displayName, identity.primaryEmail, identity.avatarUrl, now, now),
-      this.db
-        .prepare(
-          `INSERT INTO user_identities (
-             id, user_id, provider, provider_issuer, provider_user_id,
-             provider_login, provider_email, created_at
-           ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)`
-        )
-        .bind(
-          providerIdentityId,
-          userId,
-          identity.provider,
-          identity.issuer,
-          identity.subject,
-          identity.login,
-          identity.primaryEmail,
-          now
-        ),
-    ];
-    if (identity.verifiedEmails.length > 0) {
-      statements.push(
-        this.db
-          .prepare(
-            `INSERT INTO verified_email_claims (
-               email, user_id, source_kind, source_provider_identity_id,
-               created_at, last_verified_at
-             )
-             SELECT CAST(value AS TEXT), ?, 'provider_verified', ?, ?, ?
-             FROM json_each(?)`
-          )
-          .bind(userId, providerIdentityId, now, now, JSON.stringify(identity.verifiedEmails))
-      );
-    }
-    if (credential) {
-      statements.push(
-        await this.dependencies.providerCredentialStore.prepareInitialInsert(
-          providerIdentityId,
-          credential,
-          now
-        )
-      );
-    }
-
-    await this.db.batch(statements);
-
-    return {
-      userId,
-      providerIdentityId,
-      isNewUser: true,
-      collisionCount: 0,
-    };
-  }
-
-  private async findIdentity(issuer: string, subject: string): Promise<IdentityRow | null> {
-    const row = await this.db
-      .prepare(
-        `SELECT id, user_id, provider
-         FROM user_identities
-         WHERE provider_issuer = ? AND provider_user_id = ?`
-      )
-      .bind(issuer, subject)
-      .first<IdentityRow>();
-    if (!row) return null;
-    if (
-      typeof row.id !== "string" ||
-      typeof row.user_id !== "string" ||
-      typeof row.provider !== "string"
-    ) {
-      throw new Error("Stored provider identity is corrupt");
-    }
-    return row;
-  }
-
-  private async refreshExisting(
-    existing: IdentityRow,
-    identity: NormalizedIdentityEvidence,
-    credential: ProviderCredentialInput | null
-  ): Promise<ResolvedImmutableProviderIdentity> {
-    if (existing.provider !== identity.provider) {
-      throw new ProviderIdentityAdapterMismatchError();
-    }
-
-    const now = this.dependencies.clock.now();
-    const statements: SqlStatement[] = [
-      this.db
-        .prepare(
-          `UPDATE user_identities
-           SET provider_login = ?, provider_email = ?
-           WHERE id = ? AND user_id = ?`
-        )
-        .bind(identity.login, identity.primaryEmail, existing.id, existing.user_id),
-      // users.email is stable canonical account metadata, not a mirror of a
-      // provider's mutable primary email. Current provider display metadata
-      // lives on user_identities; verified ownership evidence lives in claims.
-      this.db
-        .prepare(
-          `UPDATE users
-           SET display_name = ?, avatar_url = ?, updated_at = ?
-           WHERE id = ?`
-        )
-        .bind(identity.displayName, identity.avatarUrl, now, existing.user_id),
-    ];
-    if (identity.verifiedEmails.length > 0) {
-      const serializedEmails = JSON.stringify(identity.verifiedEmails);
-      statements.push(
-        this.db
-          .prepare(
-            `UPDATE verified_email_claims
-             SET last_verified_at = ?
-             WHERE user_id = ?
-               AND source_kind != 'legacy_canonical'
-               AND email IN (SELECT CAST(value AS TEXT) FROM json_each(?))`
-          )
-          .bind(now, existing.user_id, serializedEmails),
-        this.db
-          .prepare(
-            `INSERT OR IGNORE INTO verified_email_claims (
-               email, user_id, source_kind, source_provider_identity_id,
-               created_at, last_verified_at
-             )
-             SELECT CAST(value AS TEXT), ?, 'provider_verified', ?, ?, ?
-             FROM json_each(?)`
-          )
-          .bind(existing.user_id, existing.id, now, now, serializedEmails)
-      );
-    }
-    if (credential) {
-      statements.push(
-        await this.dependencies.providerCredentialStore.prepareSignInUpsert(
-          existing.id,
-          credential,
-          now
-        )
-      );
-    }
-
-    await this.db.batch(statements);
-    return {
-      userId: existing.user_id,
-      providerIdentityId: existing.id,
-      isNewUser: false,
-      collisionCount: (await this.findConflictingEmails(identity.verifiedEmails, existing.user_id))
-        .length,
-    };
-  }
-
-  private async findClaims(emails: readonly string[]): Promise<EmailClaimRow[]> {
-    if (emails.length === 0) return [];
-    const result = await this.db
-      .prepare(
-        `SELECT email, user_id, source_kind
-         FROM verified_email_claims
-         WHERE email IN (SELECT CAST(value AS TEXT) FROM json_each(?))`
-      )
-      .bind(JSON.stringify(emails))
-      .all<EmailClaimRow>();
-
-    return result.results.map((row) => {
-      if (
-        typeof row.email !== "string" ||
-        typeof row.user_id !== "string" ||
-        (row.source_kind !== "legacy_canonical" &&
-          row.source_kind !== "provider_verified" &&
-          row.source_kind !== "trusted_bot_attribution")
-      ) {
-        throw new Error("Stored verified email claim is corrupt");
-      }
-      return row;
-    });
-  }
-
-  private async findConflictingEmails(
-    emails: readonly string[],
-    expectedUserId: string | null
-  ): Promise<string[]> {
-    const claims = await this.findClaims(emails);
-    return claims
-      .filter((claim) => expectedUserId === null || claim.user_id !== expectedUserId)
-      .map((claim) => claim.email)
-      .sort();
-  }
-}
